bug fixes
Fixed bug in pixel art editor event handlers that caused a nudge of the drawing to also move the object in the main canvas

Fixed a bug when overwriting an effect as an admin, where the admin name was saved as the developer's name instead of the original name. This is needed to fix some effects with legacy data in the database.

Other bug fixes.